pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "starrocks-br"
version = "0.1.0"
description = "StarRocks Backup and Restore automation tool"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.7,<9",
    "PyYAML>=6.0.1,<7",
    "mysql-connector-python>=9.0.0,<10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2,<9",
    "pytest-mock>=3.14.0,<4",
    "pytest-cov>=5.0.0,<6",
]

[project.scripts]
starrocks-br = "starrocks_br.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q --cov=src/starrocks_br --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src/starrocks_br"]

[tool.coverage.report]
omit = ["*/__init__.py"]
show_missing = true
skip_covered = false

[tool.black]
line-length = 100
